Revert "Fix #1118 Setup zp_session in admin toolbox so gallery delete links and their XSRF token work without GALLERY_SESSION being required to be enabled (This is what the backend itself does anyway); add "edit category" link to Zenpage toolbox links"
This reverts commit f74d5e331655ead3976bd29fda2080de81c24468.

+++ zp-core/template-functions.php
@@ -55,7 +55,6 @@ function printZenJavascripts() {
 function adminToolbox() {
 	global $_zp_current_album, $_zp_current_image, $_zp_current_search, $_zp_gallery_page, $_zp_gallery, $_zp_current_admin_obj, $_zp_loggedin, $_zp_conf_vars;
 	if (zp_loggedin()) {
-		zp_session_start();
 		$zf = FULLWEBPATH . "/" . ZENFOLDER;
 		$page = getCurrentPage();
 		ob_start();
@@ -117,6 +116,7 @@ function adminToolbox() {
 					<?php
 				}
 				zp_apply_filter('admin_toolbox_global', $zf);
+
 				if (zp_loggedin(TAGS_RIGHTS)) {
 					?>
 					<li>
@@ -176,11 +176,13 @@ function adminToolbox() {
 						}
 						if (zp_loggedin(UPLOAD_RIGHTS)) {
 							// admin has upload rights, provide an upload link for a new album
-							?>
-							<li>
-								<a href="javascript:newAlbum('',true);"><?php echo gettext("New Album"); ?></a>
-							</li>
-							<?php
+							if (GALLERY_SESSION) { // XSRF defense requires sessions
+								?>
+								<li>
+									<a href="javascript:newAlbum('',true);"><?php echo gettext("New Album"); ?></a>
+								</li>
+								<?php
+							}
 						}
 						if ($_zp_gallery_page == 'index.php') {
 							$redirect = '';
@@ -221,12 +223,15 @@ function adminToolbox() {
 								}
 							}
 							// and a delete link
-							?>
-							<li>
-								<a href="javascript:confirmDeleteAlbum('<?php echo $zf; ?>/admin-edit.php?page=edit&amp;action=deletealbum&amp;album=<?php echo urlencode(pathurlencode($albumname)) ?>&amp;XSRFToken=<?php echo getXSRFToken('delete'); ?>');"
-									 title="<?php echo gettext('Delete the album'); ?>"><?php echo gettext('Delete album'); ?></a>
-							</li>
-							<?php
+							if (GALLERY_SESSION) { // XSRF defense requires sessions
+								?>
+								<li>
+									<a href="javascript:confirmDeleteAlbum('<?php echo $zf; ?>/admin-edit.php?page=edit&amp;action=deletealbum&amp;album=<?php echo urlencode(pathurlencode($albumname)) ?>&amp;XSRFToken=<?php echo getXSRFToken('delete'); ?>');"
+										 title="<?php echo gettext('Delete the album'); ?>"><?php echo gettext('Delete album'); ?></a>
+								</li>
+								<?php
+								
+							}
 						}
 						if ($_zp_current_album->isMyItem(UPLOAD_RIGHTS) && !$_zp_current_album->isDynamic()) {
 							// provide an album upload link if the admin has upload rights for this album and it is not a dynamic album
@@ -234,10 +239,14 @@ function adminToolbox() {
 							<li>
 								<?php printLinkHTML($zf . '/admin-upload.php?album=' . pathurlencode($albumname), gettext("Upload Here"), NULL, NULL, NULL); ?>
 							</li>
-							<li>
-								<a href="javascript:newAlbum('<?php echo pathurlencode($albumname); ?>',true);"><?php echo gettext("New Album Here"); ?></a>
-							</li>
 							<?php
+							if (GALLERY_SESSION) { // XSRF defense requires sessions
+								?>
+								<li>
+									<a href="javascript:newAlbum('<?php echo pathurlencode($albumname); ?>',true);"><?php echo gettext("New Album Here"); ?></a>
+								</li>
+								<?php
+							}
 						}
 						zp_apply_filter('admin_toolbox_album', $albumname, $zf);
 						if ($inImage) {
@@ -246,11 +255,15 @@ function adminToolbox() {
 							if (!$_zp_current_album->isDynamic()) { // don't provide links when it is a dynamic album
 								if ($_zp_current_album->isMyItem(ALBUM_RIGHTS)) {
 									// if admin has edit rights on this album, provide a delete link for the image.
+									if (GALLERY_SESSION) { // XSRF defense requires sessions
+										?>
+										<li>
+											<a href="javascript:confirmDelete('<?php echo $zf; ?>/admin-edit.php?page=edit&amp;action=deleteimage&amp;album=<?php echo urlencode(pathurlencode($albumname)); ?>&amp;image=<?php echo urlencode($imagename); ?>&amp;XSRFToken=<?php echo getXSRFToken('delete'); ?>',deleteImage);"
+												 title="<?php echo gettext("Delete the image"); ?>"><?php echo gettext("Delete image"); ?></a>
+										</li>
+										<?php
+									}
 									?>
-									<li>
-										<a href="javascript:confirmDelete('<?php echo $zf; ?>/admin-edit.php?page=edit&amp;action=deleteimage&amp;album=<?php echo urlencode(pathurlencode($albumname)); ?>&amp;image=<?php echo urlencode($imagename); ?>&amp;XSRFToken=<?php echo getXSRFToken('delete'); ?>',deleteImage);"
-											 title="<?php echo gettext("Delete the image"); ?>"><?php echo gettext("Delete image"); ?></a>
-									</li>
 									<li>
 										<a href="<?php echo $zf; ?>/admin-edit.php?page=edit&amp;album=<?php echo pathurlencode($albumname); ?>&amp;singleimage=<?php echo urlencode($imagename); ?>&amp;tab=imageinfo&amp;nopagination"
 											 title="<?php echo gettext('Edit image'); ?>"><?php echo gettext('Edit image'); ?></a>
